move list_scripts function to filesystem
