+++ sabnzbd/filesystem.py
@@ -33,6 +33,7 @@ from typing import Union, List, Tuple, Any, Dict, Optional
 
 try:
     import win32api
+    import win32file
 except ImportError:
     pass
 
@@ -474,7 +475,33 @@ def fix_unix_encoding(folder: str):
 
 def is_valid_script(basename: str) -> bool:
     """ Determine if 'basename' is a valid script """
-    return basename in sabnzbd.api.list_scripts(default=False, none=False)
+    return basename in list_scripts(default=False, none=False)
+
+
+def list_scripts(default: bool = False, none: bool = True) -> List[str]:
+    """ Return a list of script names, optionally with 'Default' added """
+    lst = []
+    path = sabnzbd.cfg.script_dir.get_path()
+    if path and os.access(path, os.R_OK):
+        for script in globber_full(path):
+            if os.path.isfile(script):
+                if (
+                    (
+                        sabnzbd.WIN32
+                        and os.path.splitext(script)[1].lower() in PATHEXT
+                        and not win32api.GetFileAttributes(script) & win32file.FILE_ATTRIBUTE_HIDDEN
+                    )
+                    or script.endswith(".py")
+                    or (not sabnzbd.WIN32 and userxbit(script) and not os.path.basename(script).startswith("."))
+                ):
+                    lst.append(os.path.basename(script))
+            # Make sure capitalization is ignored to avoid strange results
+            lst = sorted(lst, key=str.casefold)
+        if none:
+            lst.insert(0, "None")
+        if default:
+            lst.insert(0, "Default")
+    return lst
 
 
 def make_script_path(script: str) -> Optional[str]:
